+++ smart-contracts/osmosis/contracts/cl-vault/src/vault/mod.rs
@@ -1,9 +1,8 @@
 pub mod admin;
-pub mod any_deposit;
 pub mod autocompound;
 pub mod concentrated_liquidity;
+pub mod deposit;
 pub mod distribution;
-pub mod exact_deposit;
 pub mod merge;
 pub mod range;
 pub mod swap;
